requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
httpx>=0.27.0
mlx>=0.18.0
mlx-lm>=0.18.0
mlx-vlm>=0.1.0
